+++ BackEnd/Backend/Backend.Common/Models/Stocks/RealTimeStock.cs
@@ -5,23 +5,23 @@ namespace Backend.Common.Models.Stocks;
 public class RealTimeStock
 {
     [JsonPropertyName("c")]
-    public double CurrentPrice { get; set; }
+    public decimal CurrentPrice { get; set; }
 
     [JsonPropertyName("d")]
-    public double Change { get; set; }
+    public decimal Change { get; set; }
 
     [JsonPropertyName("dp")]
-    public double ChangePercent { get; set; }
+    public decimal ChangePercent { get; set; }
 
     [JsonPropertyName("h")]
-    public double DayHighPrice { get; set; }
+    public decimal DayHighPrice { get; set; }
 
     [JsonPropertyName("l")]
-    public double DayLowPrice { get; set; }
+    public decimal DayLowPrice { get; set; }
     
     [JsonPropertyName("o")]
-    public double DayOpenPrice { get; set; }
+    public decimal DayOpenPrice { get; set; }
 
     [JsonPropertyName("pc")]
-    public double PreviousClosePrice { get; set; }
+    public decimal PreviousClosePrice { get; set; }
 }
