+++ BackEnd/Backend/Backend.Common/Models/Stocks/StockDailyData.cs
@@ -2,13 +2,13 @@ namespace Backend.Common.Models.Stocks;
 
 public class StockDailyData
 {
-    public double OpenPrice { get; set; }
+    public decimal OpenPrice { get; set; }
     
-    public double HighPrice { get; set; }
+    public decimal HighPrice { get; set; }
     
-    public double LowPrice { get; set; }
+    public decimal LowPrice { get; set; }
     
-    public double ClosePrice { get; set; }
+    public decimal ClosePrice { get; set; }
     
-    public double Volume { get; set; }
+    public decimal Volume { get; set; }
 }
