+++ BackEnd/Backend/Backend.InvestingAdvisor/Models/StockClassificationData.cs
@@ -2,13 +2,13 @@ namespace Backend.InvestingAdvisor.Models;
 
 public class StockClassificationData
 {
-    public double Beta { get; set; }
+    public decimal Beta { get; set; }
     
-    public double Roe { get; set; }
+    public decimal Roe { get; set; }
 
-    public double PeRatio { get; set; }
+    public decimal PeRatio { get; set; }
 
-    public double PbRatio { get; set; }
+    public decimal PbRatio { get; set; }
     
-    public double DividendYield { get; set; }
+    public decimal DividendYield { get; set; }
 }
